split CSP rbac
